refactor(): 大幅重构了Market业务，购买资源现变更为购买资源权限; 重构了Market上架、下架、审核的逻辑；调整了鉴权相关方法以从Market组获得资源权限；调整了资源级权限覆盖的相关逻辑，现改为可按组覆盖；调整了Market组新建标签的逻辑；进行了其他规范性调整

+++ wisepen-resource-service/wisepen-resource-api/src/main/java/com/oriole/wisepen/resource/domain/mq/ResourceForkMessage.java
@@ -1,6 +1,5 @@
 package com.oriole.wisepen.resource.domain.mq;
 
-import com.oriole.wisepen.resource.enums.MarketPurchaseType;
 import com.oriole.wisepen.resource.enums.ResourceType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,6 +21,6 @@ public class ResourceForkMessage {
 
     private Long forkedResourceOwnerId;
 
-    private Long forkedResourceVersion;
+    private Integer forkedResourceVersion;
 
 }
